responsive styles
setup responsive styles for screen orientation changes
