+++ src/components/member/ContainerItem.tsx
@@ -4,7 +4,7 @@ import { GestureDetector, Gesture } from "react-native-gesture-handler";
 import { ContainerObj, EquipmentObj } from "../../types/ModelTypes";
 import { useEquipment } from "../../helper/context/EquipmentContext";
 import { chunkArray } from "../../helper/EquipmentUtils";
-import { useResponsiveStyles } from "../../styles/ResponsiveStyles";
+import { useItemStyles } from "../../styles/ItemStyles";
 import EquipmentDisplay from "./EquipmentDisplay";
 
 /*
@@ -23,10 +23,10 @@ export default function ContainerItem({
 }) {
   const { setContainerItem, setContainerVisible, setSwapContainerVisible } =
     useEquipment();
+  const itemStyles = useItemStyles();
   // only display the first 9 equipment items in a 3x3 grid
   const firstNine = item?.equipment.slice(0, 9);
   const chunkedData = chunkArray(firstNine ? firstNine : [], 3);
-  const itemStyles = useResponsiveStyles();
 
   // swapable is a boolean that determines if the container opens the swapContainerOverlay
   // since swapContainerOverlay is only used in the swapEquipment screen
