+++ src/screens/organization/MemberProfileScreen.tsx
@@ -1,6 +1,5 @@
 import { Alert, Text, View } from "react-native";
 import { MemberProfileScreenProps } from "../../types/ScreenTypes";
-import { profileStyles } from "../../styles/ProfileStyles";
 import ProfileDisplay from "../../components/ProfileDisplay";
 import { Button } from "@rneui/themed";
 import { useLoad } from "../../helper/context/LoadingContext";
@@ -8,6 +7,7 @@ import { DataStore } from "@aws-amplify/datastore";
 import { Organization, OrgUserStorage } from "../../models";
 import { handleError } from "../../helper/Utils";
 import { useUser } from "../../helper/context/UserContext";
+import { useProfileStyles } from "../../styles/ProfileStyles";
 
 export default function MemberProfileScreen({
   route,
@@ -15,6 +15,7 @@ export default function MemberProfileScreen({
   const { member } = route.params;
   const { setIsLoading } = useLoad();
   const { org, setOrg, isManager, setIsManager } = useUser();
+  const profileStyles = useProfileStyles();
 
   // delete an orgUserStorage associated with the user
   // DOING SO ALSO REMOVES ALL EQUIPMENT ASSOCIATED WITH THE USER
